reorganize files, add new function for easy rebalancing
