+++ src/main/java/spireCafe/interactables/patrons/powerelic/implementation/PowerelicUpdateEnergyAndHandsizeAction.java
@@ -1,4 +1,4 @@
-package spireCafe.interactables.patrons.powerelic;
+package spireCafe.interactables.patrons.powerelic.implementation;
 
 import com.megacrit.cardcrawl.actions.AbstractGameAction;
 import com.megacrit.cardcrawl.actions.common.DrawCardAction;
