+++ src/main/java/spireCafe/interactables/patrons/powerelic/implementation/debug/PowerelicDebugRemovalEffect.java
